Refactor code: add middlewares.js file
