+++ server/src/index.js
@@ -1,30 +1,13 @@
 import express from 'express';
-import bodyParser from 'body-parser';
-import { graphiqlExpress, graphqlExpress } from 'apollo-server-express';
-import { makeExecutableSchema } from 'graphql-tools';
 
 import './config/db';
 import constants from './config/constants';
-import typeDefs from './graphql/schema';
-import resolvers from './graphql/resolvers';
 import mocks from './mocks';
+import middlewares from './config/middlewares';
 
 const app = express();
 
-const schema = makeExecutableSchema({
-  typeDefs,
-  resolvers,
-});
-
-app.use(bodyParser.json());
-
-app.use('/graphiql', graphiqlExpress({
-  endpointURL: constants.GRAPHQL_PATH,
-}));
-
-app.use(constants.GRAPHQL_PATH, graphqlExpress({
-  schema,
-}));
+middlewares(app);
 
 mocks().then(() => {
   app.listen(constants.PORT, (err) => {
